fields: Move to subpackage
